Adds support for setting custom priority
While before we were hardcoding the LOCAL0 facility and
EMERG severity to specify the priority of the messages,
after this commit the user is able to specify those via
CLI flags and have that set in the syslog message.

The commit also adds test for such funcionality (see the
`priority` package).